requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
idna==3.10
pillow==11.2.1
python-nmap==0.7.1
reportlab==4.4.0
requests==2.32.3
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
